Sakila DVD Rental database
In this project, you will use SQL to explore a database related to movie rentals. You will write SQL code to run SQL queries and answer interesting questions about the database. As part of your project submission, you will run SQL queries and build visualizations to showcase the output of your queries.

We will begin by getting familiar with the database. We have included Practice Quizzes that include a series of questions that will assure you have mastered the main concepts taught throughout the SQL lessons. These practice quizzes will not be "graded" by a reviewer, but they will help you self-assess and make sure you are on the right track. The quizzes are there to assist you in understanding the database before developing the questions that you wish to include for the project.

The project submission is a presentation, which will be reviewed, and for which you will need to Meet Expectations to pass. For the presentation component, you will create four slides. Each slide will:

Have a question of interest.
Have a supporting SQL query needed to answer the question.
Have a supporting visualization created using the final data of your SQL query that answers your question of interest.
You will submit your project at the end of the project lessons. Your project will include:

A set of slides with a question, visualization, and small summary on each slide.
A text file with your queries needed to answer each of the four questions.
The essentials of your project submission are discussed on the page labeled as Project Submission.

Additional Resources
For the second part of this project, you may want a refresher of statistics or working with spreadsheets (perhaps specific to creating data visualizations). There are additional resources in the extracurriculars of this classroom to assist with these ideas if you need them!